requirements.txt
langchain>=0.1.13
langchain-community>=0.0.21
langchain-openai
openai
faiss-cpu
python-dotenv
beautifulsoup4
requests